Fix
https://github.com/patchthecode/JTAppleCalendar/issues/304

User changed frame size AND the number of rows displayed upon selection of a date.

Because he selected the 31st, it had a counterpart date of a particular index which was reloaded on an async thread.

When the rows were changed that counterpart date was no longer on the screen. So when the async reaload code was called, it tried to reload an index that didnt exist resulting in a crash.