Fixed bug.
We should check to see if the cell is indeeded selected on the calendar here.
If user reloaded the calendar in multiselect mode, the collectionView will remove all its selected cells. Our custom SelectCells array will have the cell as selected but the collection view will not. The result is the cells still looks as if it is selected, but when user taps, it becomes selected (because it wasnt selected in the eyes of the collectionView)