Updated
https://github.com/patchthecode/JTAppleCalendar/issues/1014#issuecomment-483420575

When checking oin which paths to reload, this should not be restriction to section because a user might use range selection with their own logic rather than the one provided by the library. In a case like this, we would want all forward and backward indexes inrelation that indexPath X to be updated (regardless on section).